rename & relocate hero
