+++ src/data/homeHeroPhrases.tsx
